Split internal notify endpoint into two endpoints
