+++ app/controllers/internal/v0/notify_controller.rb
@@ -1,13 +1,25 @@
 module Internal
   module V0
     class NotifyController < ::ApplicationController
-      def notify
-        klass = params.require(:klass)
+      def notify_approval_request
         id = params.require(:id)
         payload = params.require(:payload)
 
         ActsAsTenant.without_tenant do
-          Catalog::Notify.new(klass, id, payload).process
+          Catalog::NotifyApprovalRequest.new(id, payload).process
+        end
+
+        json_response(nil)
+      end
+
+      def notify_order_item
+        task_id = params.require(:task_id)
+        payload = params.require(:payload)
+        message = params.require(:message)
+
+        topic = Struct.new(:payload, :message).new(payload.merge({"task_id" => task_id}), message)
+        ActsAsTenant.without_tenant do
+          Catalog::UpdateOrderItem.new(topic).process
         end
 
         json_response(nil)
