+++ config/routes.rb
@@ -16,7 +16,8 @@ Rails.application.routes.draw do
     match "/v0/*path", :via => [:post], :to => redirect(:path => "/internal/v1.0/%{path}", :only_path => true)
 
     namespace :v1x0, :controller => 'notify', :path => "v1.0" do
-      post '/notify/:klass/:id', :action => 'notify'
+      post '/notify/approval_request/:id', :action => 'notify_approval_request'
+      post '/notify/order_item/:task_id', :action => 'notify_order_item'
     end
   end
