+++ spec/controllers/internal/v1x0/notify_controller_spec.rb
@@ -1,33 +1,42 @@
 describe Internal::V1x0::NotifyController, :type => :request do
-  describe "POST /notify/:klass/:id" do
-    around do |example|
-      bypass_rbac do
-        example.call
-      end
+  around do |example|
+    bypass_rbac do
+      example.call
+    end
+  end
+
+  describe "POST /notify/approval_request/:id" do
+    let(:approval_request) { create(:approval_request) }
+
+    it "returns a 200" do
+      post "/internal/v1.0/notify/approval_request/#{approval_request.id}", :headers => default_headers, :params => {:payload => {:decision => "test"}.to_json}
+
+      expect(response.status).to eq(200)
+    end
+  end
+
+  describe "POST /notify/order_item/:task_id" do
+    let(:tenant) { create(:tenant) }
+    let(:order) { create(:order, :tenant_id => tenant.id) }
+    let(:portfolio_item) { create(:portfolio_item, :service_offering_ref => "123", :tenant_id => tenant.id) }
+    let!(:order_item) do
+      create(:order_item,
+             :order_id          => order.id,
+             :portfolio_item_id => portfolio_item.id,
+             :tenant_id         => tenant.id,
+             :topology_task_ref => "321")
+    end
+    let(:update_order_item) { instance_double("Catalog::UpdateOrderItem") }
+
+    before do
+      allow(Catalog::UpdateOrderItem).to receive(:new).and_return(update_order_item)
+      allow(update_order_item).to receive(:process)
     end
 
-    context "when the class provided is not a supported notification class" do
-      let(:klass) { "portfolio" }
+    it "returns a 200" do
+      post "/internal/v1.0/notify/order_item/321", :headers => default_headers, :params => {:payload => {:status => "test"}, :message => "message"}
 
-      it "returns a 422" do
-        post "/internal/v1.0/notify/#{klass}/123", :headers => default_headers, :params => {:payload => {:decision => "test"}.to_json}
-
-        expect(response.status).to eq(422)
-      end
-    end
-
-    context "when the class provided is supported" do
-      let(:klass) { "order_item" }
-      let(:tenant) { create(:tenant) }
-      let(:order) { create(:order, :tenant_id => tenant.id) }
-      let(:portfolio_item) { create(:portfolio_item, :service_offering_ref => "123", :tenant_id => tenant.id) }
-      let!(:order_item) { create(:order_item, :order_id => order.id, :portfolio_item_id => portfolio_item.id, :tenant_id => tenant.id) }
-
-      it "returns a 200" do
-        post "/internal/v1.0/notify/#{klass}/#{order_item.id}", :headers => default_headers, :params => {:payload => {:decision => "test"}.to_json}
-
-        expect(response.status).to eq(200)
-      end
+      expect(response.status).to eq(200)
     end
   end
 end
